프로젝트 구조를 MVC 패턴으로 변경
