--- models/Product.js
@@ -1,30 +1,0 @@
-// Product 스키마
-import mongoose from "mongoose";
-
-const productSchema = new mongoose.Schema(
-  {
-    name: {
-      type: String,
-      required: true,
-    },
-    description: {
-      type: String,
-      required: true,
-    },
-    price: {
-      type: Number,
-      required: true,
-    },
-    tags: {
-      type: [String],
-      default: [],
-    },
-  },
-  {
-    timestamps: true,
-  },
-);
-
-const Product = mongoose.model("Product", productSchema);
-
-export default Product;
